requirements.txt
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
emoji==2.12.1
fonttools==4.53.0
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
idna==3.7
imageio==2.34.2
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
multitasking==0.0.11
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.35.0
tenacity==8.4.1
tifffile==2024.6.18
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
webencodings==0.5.1
yfinance==0.2.40
